requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.27,<1.0
httpx-sse>=0.4,<1.0
qdrant-client>=1.9,<2.0
redis>=5.0,<6.0
asyncpg>=0.29,<1.0
neo4j>=5.20,<6.0
pydantic-settings>=2.3,<3.0
structlog>=24.1,<25.0
PyYAML>=6.0,<7.0
